refactor: extract text editor copy paste handling
